+++ src/test/java/gift/ProductControllerTest.java
@@ -1,11 +1,13 @@
 package gift;
 
+import gift.controller.ProductController;
+import gift.entity.Product;
+import gift.entity.ProductDao;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mockito.InjectMocks;
 import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
-import org.springframework.jdbc.core.JdbcTemplate;
 
 import java.util.HashMap;
 import java.util.List;
